Add device helper classes
This commit adds helper classes that record device information. These
classes are not yet implemented into the other parts of the project.
These classes are intended to simplify the interface between the main
project and the mock classes. These classes are not tested as currently
they are simple getter and setter classes.