Introduce Command interface and subclasses
This commit is a large change to cleanup the logic of OnboardingApp.
This commit was made to reduce the density of branching logic in
OnboardingApp. Each command now has its own class, hence OnboardingApp
only contains the logic for routing and no command logic. Therefore,
the logic for each of the commands and the OnboardingApp is now
separated better than previously. This makes it easier to add new
commands and to change the logic of existing commands.